// application/frontend/src/app/shared/models/load-demand.ts
/**
 * @license
 * Copyright 2022 Google LLC
 *
 * Use of this source code is governed by an MIT-style
 * license that can be found in the LICENSE file or at
 * https://opensource.org/licenses/MIT.
 */

import { ILoad } from '../../core/models';

export interface LoadDemandFormValue {
  type: string;
  value: ILoad;
}